requirements.txt
flake8==7.1.1
isort==5.10.1
kubernetes==29.0.0
pre-commit==2.15.0
pyfiglet==0.7
PyGithub==2.3.0
PyYAML==6.0.1
questionary==1.10.0
requests==2.27.1
rich==10.16.2
stripe==10.4.0
beautifulsoup4==4.12.3
Markdown==3.6
Pygments==2.18.0
